Get rid of small warnings insdie Hazel module

+++ Hazel/src/Platform/Windows/WindowsWindow.cpp
@@ -111,13 +111,13 @@ namespace Hazel {
 
 		glfwSetScrollCallback(m_Window, [](GLFWwindow* window, double xOffset, double yOffset) {
 			WindowData& data = *(WindowData*)glfwGetWindowUserPointer(window);
-			MouseScrolledEvent event(xOffset, yOffset);
+			MouseScrolledEvent event((float)xOffset, (float)yOffset);
 			data.EventCallback(event);
 			});
 
 		glfwSetCursorPosCallback(m_Window, [](GLFWwindow* window, double xPosition, double yPosition) {
 			WindowData& data = *(WindowData*)glfwGetWindowUserPointer(window);
-			MouseMovedEvent event(xPosition, yPosition);
+			MouseMovedEvent event((float)xPosition, (float)yPosition);
 			data.EventCallback(event);
 			});
